+++ kernel/api/sys/types.h
@@ -3,21 +3,21 @@
 #include <stdint.h>
 
 typedef uint32_t dev_t;
-typedef uint32_t uid_t;
-typedef uint32_t gid_t;
-typedef uint32_t ino_t;
-typedef uint32_t mode_t;
+typedef unsigned int uid_t;
+typedef unsigned int gid_t;
+typedef unsigned long ino_t;
+typedef unsigned int mode_t;
 typedef uint32_t nlink_t;
-typedef int32_t off_t;
-typedef int64_t loff_t;
-typedef int32_t pid_t;
-typedef uint32_t clock_t;
-typedef int64_t time_t;
-typedef uint32_t useconds_t;
-typedef int32_t suseconds_t;
-typedef int32_t blksize_t;
-typedef int32_t blkcnt_t;
-typedef int32_t ssize_t;
+typedef long off_t;
+typedef long long loff_t;
+typedef int pid_t;
+typedef long clock_t;
+typedef long time_t;
+typedef unsigned int useconds_t;
+typedef long suseconds_t;
+typedef long blksize_t;
+typedef uint64_t blkcnt_t;
+typedef long ssize_t;
 
-typedef uint16_t linux_old_uid_t;
-typedef uint16_t linux_old_gid_t;
+typedef unsigned short linux_old_uid_t;
+typedef unsigned short linux_old_gid_t;
